requirements.txt
clipboard-monitor==1.0.5
decorator==5.1.1
future==0.18.3
numpy==1.24.3
pandas==2.0.1
python-dateutil==2.8.2
python-whois==0.8.0
pytz==2023.3
pywin32==306
six==1.16.0
tzdata==2023.3
validators==0.20.0